+++ app/src/main/java/com/websarva/wings/android/abiityofbaseball/activity/AnswerQuestionsActivity.kt
@@ -7,9 +7,10 @@ import android.view.View
 import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import com.websarva.wings.android.abiityofbaseball.*
-import com.websarva.wings.android.abiityofbaseball.fragment.QuestionOfAppearanceFragment
-import com.websarva.wings.android.abiityofbaseball.fragment.QuestionOfPersonalityFragment
-import com.websarva.wings.android.abiityofbaseball.fragment.QuestionOfPopularityFragment
+import com.websarva.wings.android.abiityofbaseball.calc_ability.*
+import com.websarva.wings.android.abiityofbaseball.fragment.question.QuestionOfAppearanceFragment
+import com.websarva.wings.android.abiityofbaseball.fragment.question.QuestionOfPersonalityFragment
+import com.websarva.wings.android.abiityofbaseball.fragment.question.QuestionOfPopularityFragment
 import kotlinx.android.synthetic.main.activity_player_making.*
 import kotlinx.android.synthetic.main.fragment_question_of_appearance.*
 import kotlinx.android.synthetic.main.fragment_question_of_personality.*
