+++ app/src/main/java/com/websarva/wings/android/abiityofbaseball/fragment/player_info/FielderRecordFragment.kt
@@ -1,4 +1,4 @@
-package com.websarva.wings.android.abiityofbaseball.fragment
+package com.websarva.wings.android.abiityofbaseball.fragment.player_info
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.websarva.wings.android.abiityofbaseball.Constants
-import com.websarva.wings.android.abiityofbaseball.PlayerFielderClass
+import com.websarva.wings.android.abiityofbaseball.player_class.PlayerFielderClass
 import com.websarva.wings.android.abiityofbaseball.R
 import kotlinx.android.synthetic.main.fragment_fielder_record.*
 import kotlin.properties.Delegates
