+++ app/src/main/java/com/websarva/wings/android/abiityofbaseball/fragment/player_info/NameFragment.kt
@@ -1,4 +1,4 @@
-package com.websarva.wings.android.abiityofbaseball.fragment
+package com.websarva.wings.android.abiityofbaseball.fragment.player_info
 
 import android.os.Bundle
 import android.util.TypedValue
