+++ app/src/main/java/com/websarva/wings/android/abiityofbaseball/player_class/PlayerPitcherClass.kt
@@ -1,7 +1,7 @@
-package com.websarva.wings.android.abiityofbaseball
+package com.websarva.wings.android.abiityofbaseball.player_class
 
+import com.websarva.wings.android.abiityofbaseball.Constants
 import java.io.Serializable
-import kotlin.math.ceil
 
 class PlayerPitcherClass(
         val playerName: String,
