+++ app/src/main/res/layout/fragment_name.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.NameFragment">
+    tools:context=".fragment.player_info.NameFragment">
 
     <TextView
         android:id="@+id/name_box"
