+++ app/src/main/res/layout/fragment_player_info.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/player_info_background"
     android:orientation="vertical"
-    tools:context=".fragment.PlayerInfoFragment">
+    tools:context=".fragment.player_info.PlayerInfoFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
