+++ app/src/main/res/layout/fragment_question_of_personality.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="#ffffff"
     android:orientation="vertical"
-    tools:context=".fragment.QuestionOfAppearanceFragment">
+    tools:context=".fragment.question.QuestionOfAppearanceFragment">
 
 
     <View
